% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_Cal_val}
\alias{get_Cal_val}
\title{CPP_CalValue}
\usage{
get_Cal_val(x, y)
}
\arguments{
\item{x}{matrix of independent variables}

\item{y}{matrix of dependent variables}
}
\value{
beta a vector of estimated coefficients

fitted_val a vector of estimated fitted value

resid_val a vector of estimated residuals

xtx a matrix calculate (X^T*X)^(-1)

df a double value represents degree of freedom
}
\description{
Use RCPPArmadillo to speed up calculation for 
inside function, do not need to export
}
