% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R
\name{lm2}
\alias{lm2}
\title{lm2}
\usage{
lm2(formula, data, na.handle = "na.omit", res_display = TRUE)
}
\arguments{
\item{formula}{An object of class "formula" that models regression to be fit}

\item{data}{Data that perform the linear regression on
#'}

\item{na.handle}{Opition to deal with NA. Two options valid: "na.omit"-- omit NA in data, and "na.fail" -- Report an error and stop the function}

\item{res_display}{Bool value that used to contorl display lm2 result or not}
}
\value{
a list that contains the following values: coefficients, call, x, y, Missing.N
}
\description{
lm2 is used to fit linear models with regression
}
\examples{
lm2(formula = fixed.acidity ~ volatile.acidity + citric.acid * residual.sugar, data = wine)
example <- lm2(formula = quality ~ free.sulfur.dioxide:pH + pH, data = wine)
example$coefficients

}
