\name{lm2Wsmry-package}
\alias{lm2Wsmry-package}
\alias{lm2Wsmry}
\docType{package}
\title{
\packageTitle{lm2Wsmry}
}
\description{
\packageDescription{lm2Wsmry}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lm2Wsmry}
\packageIndices{lm2Wsmry}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lm2Wsmry}

Maintainer: \packageMaintainer{lm2Wsmry}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
