% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_lm2.R
\name{summary_lm2}
\alias{summary_lm2}
\title{summary_lm2}
\usage{
summary_lm2(lm_mod, res_display = TRUE)
}
\arguments{
\item{lm_mod}{output value of lm2() in 'lm2Wsmry'.}

\item{res_display}{Bool value that used to contorl display summary_lm2 result or not. The default value is set to TRUE. Set it to FALSE if displaying is not required.}
}
\value{
a list that contains the following values: call(formula), residuals, coefficients, fitted.values, RSE, df(degree of freedom), r.squared, adj.r.squared, F statistic, f.pval(p-value), missing.N(missing numbers of data due to NA), cov.unscaled(unscaled covarience), rank, resd.tb(residual quntile table), coef.tb(coefficient table with significant level).
}
\description{
summary_lm2 is used to summarize fit for linear regression models.
}
\examples{
summary_lm2(lm2(formula = Rape ~ Murder+Assault+UrbanPop, data = USArrests, res_display = FALSE))


}
