% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R
\name{lm2}
\alias{lm2}
\title{lm2}
\usage{
lm2(formula, data, na.handle = "na.omit", res_display = TRUE)
}
\arguments{
\item{formula}{An object of class "formula" that models regression to be fit.}

\item{data}{Data that perform the linear regression on.}

\item{na.handle}{Opition to deal with NA. Two options valid: "na.omit"-- omit NA in data, and "na.fail" -- Report an error and stop the function. The default value is set to "na.omit".}

\item{res_display}{Bool value that used to contorl display lm2 result or not. The default value is set to TRUE. Set it to FALSE if displaying is not required.}
}
\value{
A list that contains the following values: call(formula), coefficients, residuals, fitted.values, x(independent variables), y(dependent variable),missing.N(missing numbers of data due to NA), df(residual degree of freedom) and betas(coefficient in other format)
}
\description{
lm2 is used to fit linear models with regression.
}
\examples{
lm2(formula = Temp ~ Wind + Solar.R, data = airquality, na.handle = "na.omit")

result = lm2(formula = Rape ~ Murder+Assault+UrbanPop, data = USArrests, res_display = FALSE)
result$coefficients

}
