% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_lm2.R
\name{summary_lm2}
\alias{summary_lm2}
\title{summary_lm2}
\usage{
summary_lm2(lm_mod)
}
\arguments{
\item{lm_mod}{return of lm2.}
}
\value{
a list that contains the following values: call, residuals, coefficients, RSE, df.residual, r.squared, adj.r.squared, fstatistic, f.pval, missing.N, cov.unscaled
}
\description{
summary_lm2 is used to summarize fit for linear regression models
}
\examples{

}
